Add order chats to admin dashboard
